feat: rename tissue type to sample tumor status
