+++ README.md
@@ -45,15 +45,15 @@ This demo transformer has been created to transform the values of `boy` to the C
 Another example follows for transforming Swedish spelling of `Normal`:
 
 ```php
-class AcmeTissueTypeTransformer extends TissueTypeTransformer
+class AcmeSampleTumorStatusTransformer extends SampleTumorStatusTransformer
 {
     public static function getMappings(): array
     {
-        $mappings = TissueTypeTransformer::$MAPPINGS;
+        $mappings = SampleTumorStatusTransformer::$MAPPINGS;
 
         $mappings[] =
             [
-                'value' => TissueType::NORMAL,
+                'value' => SampleTumorStatus::NORMAL,
                 'regex' => '/^(vanligt)$/i'
             ];
